+++ src/D20Tek.BlazorComponents.Tiles/LayoutOption.cs
@@ -1,0 +1,8 @@
+namespace D20Tek.BlazorComponents;
+
+public enum LayoutOption
+{
+    Compact = 0,
+    Common = 1,
+    Verbose = 2,
+}