+++ src/D20Tek.BlazorComponents.Tiles/Tile.razor.cs
@@ -1,0 +1,5 @@
+namespace D20Tek.BlazorComponents;
+
+public partial class Tile : TileBase
+{
+}