+++ src/D20Tek.BlazorComponents.Tiles/TileLayoutMetadata.cs
@@ -1,0 +1,13 @@
+namespace D20Tek.BlazorComponents;
+
+internal class TileLayoutMetadata
+{
+    private static readonly Dictionary<LayoutOption, string> _elements = new()
+    {
+        { LayoutOption.Compact, "tile-layout-compact" },
+        { LayoutOption.Common, "tile-layout-common" },
+        { LayoutOption.Verbose, "tile-layout-verbose" },
+    };
+
+    public static string GetLayoutCss(LayoutOption layout) => _elements[layout];
+}