+++ tests/D20Tek.BlazorComponents.UnitTests/GlobalUsings.cs
@@ -2,6 +2,7 @@
 global using D20Tek.Functional;
 global using Microsoft.AspNetCore.Components;
 global using Microsoft.AspNetCore.Components.Forms;
+global using Microsoft.AspNetCore.Components.Web;
 global using Microsoft.Extensions.DependencyInjection;
 global using Microsoft.VisualStudio.TestTools.UnitTesting;
 global using System;